fix(ui): added deletion for handle on subworkflow
